requirements.txt
anyio==3.6.1
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
cryptography==37.0.4
ecdsa==0.18.0
fastapi==0.109.0
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==2.0.25
starlette==0.35.0
typing_extensions==4.8.0
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3
pymysql==1.0.2
yeelight==0.7.13
apscheduler==3.10.4